Move contracts to application layer;
