+++ Application/include/Windows/IWindow.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-namespace PixelPad::Infrastructure
+namespace PixelPad::Application
 {
 	class IWindow
 	{
